+++ hutool-core/src/main/java/cn/hutool/core/convert/Convert.java
@@ -982,12 +982,27 @@ public class Convert {
 	 * 将阿拉伯数字转为中文表达方式
 	 *
 	 * @param number 数字
-	 * @param isUseTraditonal 是否使用繁体字（金额形式）
+	 * @param isUseTraditional 是否使用繁体字（金额形式）
 	 * @return 中文
 	 * @since 3.2.3
 	 */
-	public static String numberToChinese(double number, boolean isUseTraditonal) {
-		return NumberChineseFormatter.format(number, isUseTraditonal);
+	public static String numberToChinese(double number, boolean isUseTraditional) {
+		return NumberChineseFormatter.format(number, isUseTraditional);
+	}
+
+	/**
+	 * 数字中文表示形式转数字
+	 * <ul>
+	 *     <li>一百一十二 -》 112</li>
+	 *     <li>一千零一十二 -》 1012</li>
+	 * </ul>
+	 *
+	 * @param number 数字中文表示
+	 * @return 数字
+	 * @since 5.6.0
+	 */
+	public static int chineseToNumber(String number){
+		return NumberChineseFormatter.chineseToNumber(number);
 	}
 
 	/**